requirements.txt
Django~=3.0.7
djangorestframework~=3.12.4
psycopg2~=2.8.6
coverage~=5.5